Initial crack at "Legend" design
Uses the "legend" background image and simplifies the header
markup and CSS a bit. The header background colors are pulled
out of the fedora logo so things flow smoothly.

I was going for something a bit more edgy. As of now, the
design feels kind of soft. Perhaps we can spice it up later.